% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRnova.R
\name{generalizedLinearModel_}
\alias{generalizedLinearModel_}
\title{Generalized Linear Model}
\usage{
generalizedLinearModel_()
}
\description{
This is a minor modification of \code{\link[Rcmdr]{generalizedLinearModel}}
where size effects are computed and displayed for logistic regression
}
\seealso{
\code{\link[Rcmdr]{generalizedLinearModel}}
}
