% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRnova.R
\name{multiWayAnova_}
\alias{multiWayAnova_}
\title{Multiway ANOVA}
\usage{
multiWayAnova_()
}
\value{
None
}
\description{
This is a modification of \code{Rcmdr::multiWayAnova()}
where supplementary options have  been added.
}
\details{
Options:  
\itemize{
\item{\code{'SS type'}: type of sum of squared, default: \code{type = 2}. 
See Details in \code{\link[car]{Anova}}}
\item{\code{'Effect size'}: compute and prints effect size (partial eta squares)}
\item{\code{'Summary statistics for groups'}: prints summary statistics for
groups formed by all combinatuions of factors}
\item{\code{'Pairwise comparisons of means'}: performs post-hoc Tukey's HSD test  
on significant (p < .05) or close to significant (p < 0.1) effects.}
}  
On OK, the following operations are carried out:
\itemize{
 \item{Computes ANOVA using \code{\link[car]{Anova}}} 
 \item{Computes effect sizes (partial eta squared)}
 \item{Prints a summary of marginal statistics 
 (count, min, max, mean, ds)}
 \item{runs post-hoc analysis on significant or close to significant effects}
 \item {Generates an 'extended' dataset (extension  \code{.ext}) containing
 additionak columns \code{'<factorA.factorB:...>'} that allows differentiate
 measures from groups or subjects with same factors levels.
 This 'extended' dataset is useful for ploting means and post-hoc analysis  
 }
}
}
\seealso{
\code{\link[car]{Anova}} for the computation of ANOVA
}
