% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aRnova.R
\name{repMeasAnovaSetup}
\alias{repMeasAnovaSetup}
\title{Repeated measure ANOVA setup}
\usage{
repMeasAnovaSetup()
}
\description{
Dialog box to enter the names and levels of within-factors.
}
\details{
Up to three factors can be entered. A valid within-factor entry must 
consist in a syntactically valid name (see \link{make.names}) and 2 
levels or more.

On OK:
\itemize{
 \item {} {The first valid entries are kept and stored in 
 \code{.withinfactors} and \code{.withinlevels} for factor names and 
 levels, respectively.} 
 \item {} {The next dialog box (\code{repMeasAnova(.withinfactors, .withinfactors)}
  is launched.
 }
}
}
\seealso{
\link{repMeasAnova}.
}
\author{
Jessica Mange \email{jessica.mange@unicaen.fr}

Arnaud Travert \email{arnaud.travert@unicaen.fr}
}
